+++ refinedstorage-quartz-arsenal-neoforge/src/main/java/com/refinedmods/refinedstorage/quartzarsenal/neoforge/package-info.java
@@ -1,4 +1,7 @@
 @ParametersAreNonnullByDefault
+@FieldsAndMethodsAreNonnullByDefault
 package com.refinedmods.refinedstorage.quartzarsenal.neoforge;
+
+import com.refinedmods.refinedstorage.api.core.FieldsAndMethodsAreNonnullByDefault;
 
 import javax.annotation.ParametersAreNonnullByDefault;