+++ CHANGELOG.md
@@ -6,3 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Wireless Crafting Grid
+
+### Changed
+
+- You can now configure the energy usage of the Wireless Crafting Grid's recipe transfer and autocrafting.
+- You can now use the Wireless Crafting Grid without being connected to a network.